requirements.txt.txt
streamlit
openai
Pillow
requests
python-dotenv